requirements.txt
flask~=1.1.2
pyyaml~=5.3.1
requests~=2.25.1
google-api-python-client
google-auth-httplib2
google-auth-oauthlib